tests(dsl): Use snapshot testing for BraneScript and Bakery
Also moved these integration tests to the correct location for
integration tests
