+++ brane-dsl/src/compiler.rs
@@ -26,71 +26,6 @@ use crate::scanner::{self, Span, Token, Tokens};
 use crate::spec::Language;
 
 
-/***** TESTS *****/
-#[cfg(test)]
-pub mod tests {
-    use brane_shr::utilities::{create_package_index, test_on_dsl_files};
-
-    use super::*;
-
-
-    /// Tests BraneScript files.
-    #[test]
-    fn test_bscript() {
-        // Simply pass to the compiler
-        test_on_dsl_files("BraneScript", |path, code| {
-            // Print the header always
-            println!("{}", (0..80).map(|_| '-').collect::<String>());
-            println!("File '{}' gave us:", path.display());
-
-            // Read the package index
-            let pindex: PackageIndex = create_package_index();
-
-            // Create a compiler and compile it;
-            let res: Program = match parse(code, &pindex, &ParserOptions::bscript()) {
-                Ok(res) => res,
-                Err(err) => {
-                    panic!("Failed to parse BraneScript file '{}': {}", path.display(), err);
-                },
-            };
-
-            // Print it for good measure
-            println!("{res:#?}");
-            println!("{}\n\n", (0..80).map(|_| '-').collect::<String>());
-        });
-    }
-
-    /// Tests Bakery files.
-    #[test]
-    fn test_bakery() {
-        // Simply pass to the compiler
-        test_on_dsl_files("Bakery", |path, code| {
-            // Print the header always
-            println!("{}", (0..80).map(|_| '-').collect::<String>());
-            println!("File '{}' gave us:", path.display());
-
-            // Read the package index
-            let pindex: PackageIndex = create_package_index();
-
-            // Create a compiler and compile it;
-            let res: Program = match parse(code, &pindex, &ParserOptions::bakery()) {
-                Ok(res) => res,
-                Err(err) => {
-                    panic!("Failed to parse Bakery file '{}': {}", path.display(), err);
-                },
-            };
-
-            // Print it for good measure
-            println!("{res:#?}");
-            println!("{}\n\n", (0..80).map(|_| '-').collect::<String>());
-        });
-    }
-}
-
-
-
-
-
 /***** AUXILLARY STRUCTS *****/
 /// Defines options that configure the compiler before we use it.
 #[derive(Clone, Debug)]
